adding better support for rails rspec and rails dm2 format
