+++ lib/mutant.rb
@@ -99,6 +99,9 @@ require 'mutant/killer/forking'
 require 'mutant/strategy'
 require 'mutant/strategy/rspec'
 require 'mutant/strategy/rspec/example_lookup'
+require 'mutant/strategy/rspec/dm2_lookup'
+require 'mutant/strategy/rspec/rails_dm2_lookup'
+require 'mutant/strategy/rspec/rails_lookup'
 require 'mutant/runner'
 require 'mutant/cli'
 require 'mutant/color'
